backend/pyproject.toml
[project]
name = "smart-ai-support-assistant-backend"
version = "0.1.0"
description = "Smart AI Support Assistant — MindVista Technical Assessment"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "langchain-core>=1.2.23",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "uvicorn[standard]>=0.42.0",
]

